+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod app;
 pub mod config;
 pub mod event;
 pub mod export;
+pub mod import;
 pub mod index;
 pub mod link;
 pub mod markdown;